refactor: relocate tool definitions to llms package
