+++ internal/llms/hello.go
@@ -1,4 +1,4 @@
-package tools
+package llms
 
 import (
 	"context"
